Added project dir and Organized
