+++ dev/project/data/MegaCoinList.txt
