+++ dev/project/data/NegaCoinList.js
