+++ dev/project/data/RegaCoinList.js
