+++ dev/project/dataCollection.js
@@ -1,0 +1,23 @@
+import fs from 'fs'
+import { coinListURL, coinList, exchangeList } from './constants.js'
+/**
+ *This file is supposed to grab the coinLists and exchangeList and generate a file with relveant data
+ *
+ *
+ *
+ *
+ * **/
+
+function getAllCoins() {
+	fetch(coinListURL)
+		.then(response => response.json())
+		.then(data => {
+			const coinIDS = data.map(coin => coin.id)
+			const coins = data.map(coin => coin)
+			fs.writeFile("RegaCoinList.js", `const midTierCoinList = ${JSON.stringify(coins)}`.toString(), err => {
+				if (err) console.err(err)
+			})
+		})
+}
+
+getAllCoins();
