--- dev/weatherApi.json
@@ -1,1 +1,0 @@
-[object Response]
